% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_transition.R
\docType{methods}
\name{add_summaryTransition}
\alias{add_summaryTransition}
\alias{+,summaryTransition,summaryTransition-method}
\alias{subtract_summaryTransition}
\alias{-,summaryTransition,summaryTransition-method}
\title{Addition and subtraction for objects of class \code{summaryTransition}}
\usage{
add_summaryTransition(e1, e2)

\S4method{+}{summaryTransition,summaryTransition}(e1, e2)

subtract_summaryTransition(e1, e2)

\S4method{-}{summaryTransition,summaryTransition}(e1, e2)
}
\arguments{
\item{e1}{the first object}

\item{e2}{the second object}
}
\description{
Addition and subtraction for objects of class \code{summaryTransition}
}
\keyword{internal}
