% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_general_1Wrappers.R,
%   R/bout_general_2ClusterBased.R
\name{bouts_internal}
\alias{bouts_internal}
\alias{collapse_runs}
\alias{group_runs_clusterBased}
\alias{process_clusterBased_set}
\alias{check_clusterBased}
\alias{test_clusterBased_bouts}
\alias{check_no_bouts}
\title{Bout analysis internal functions}
\usage{
collapse_runs(x)

group_runs_clusterBased(
  x,
  target,
  target_buffer,
  required_percent,
  longest_allowable_interruption,
  max_n_interruptions
)

process_clusterBased_set(
  runs,
  x,
  required_percent,
  longest_allowable_interruption,
  max_n_interruptions
)

check_clusterBased(x, target = NULL, required_percent, target_buffer)

test_clusterBased_bouts(
  x,
  runs,
  tree,
  h,
  required_percent,
  max_n_interruptions,
  longest_allowable_interruption
)

check_no_bouts(results, x, target)
}
\arguments{
\item{x}{A vector of activity labels, e.g. minute-by-minute classifications
of \code{sedentary behavior}, \code{light physical activity}, or
\code{moderate-to-vigorous physical activity}. Must be an atomic vector.
Factors should be cast to integer or character prior to passing into
this function.}

\item{target}{Value of \code{x} for which bout information is desired}

\item{target_buffer}{numeric. Maximum separation between runs of
\code{target}, beyond which they will not be clustered together}

\item{required_percent}{numeric (1-100). The minimum percentage of the full
bout period that must be spent engaging in the target behavior. Stated
differently, this threshold stipulates that interruptions can compose no
more than \code{100-required_percent} of the bout}

\item{longest_allowable_interruption}{numeric. The maximum length for any
single interruption in a valid bout}

\item{max_n_interruptions}{numeric. The maximum number of interruption events
that are allowed before a bout will be considered invalid}

\item{runs}{pre-processed input (mostly run length encoded)}

\item{results}{output from \code{group_runs_clusterBased}}
}
\description{
Bout analysis internal functions
}
\keyword{internal}
