% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_reorder.R
\name{df_reorder}
\alias{df_reorder}
\alias{validate_df_reorder_input}
\title{Reorder the columns of a data frame}
\usage{
df_reorder(df, columns, after)

validate_df_reorder_input(df, arg, label)
}
\arguments{
\item{df}{the data frame}

\item{columns}{the column(s) to move (either as character names or numeric
indices)}

\item{after}{the column after which to insert \code{columns} (must be a
scalar, either a character name or a numeric index)}

\item{arg}{internal parameter (one of \code{columns} or \code{after})}

\item{label}{internal parameter (one of \code{"columns"} or \code{"after"})}
}
\value{
The reordered data frame
}
\description{
Reorder the columns of a data frame
}
\examples{
df <- data.frame(a = 1:10, b = 11:20, c = 21:30, d = 31:40)
df_reorder(df, 2:3, "d")
df_reorder(df, c("c", "d"), "a")
}
