% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent_at.R
\name{equivalent_at}
\alias{equivalent_at}
\alias{get_absolute_equivalent_at}
\title{Determine the minimum equivalence zone necessary for establishing equivalence
in a paired equivalence test}
\usage{
equivalent_at(result)

get_absolute_equivalent_at(result)
}
\arguments{
\item{result}{data frame constructed in
\code{\link{paired_equivalence_wrapper}} that provides information about
the paired equivalence test}
}
\description{
Paired equivalence tests are conducted based on a pre-specified zone of
equivalence. Following the test, it is useful to know how minimally small or
large the zone would have needed to be in order for the test to indicate
equivalence.
}
\keyword{internal}
