% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Example data for calculating bouts of moderate-to-vigorous physical activity}
\format{An object of class \code{data.frame} with 10080 rows and 12 columns.}
\usage{
ex_data
}
\description{
A dataset containing accelerometer data and predicted energy expenditure in
metabolic equivalents (METs) that can be used to classify
moderate-to-vigorous physical activity in continuous bouts.
}
\keyword{datasets}
