% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_general_1Wrappers.R
\name{get_bouts}
\alias{get_bouts}
\title{Identify bouts of activity, allowing for interruptions}
\usage{
get_bouts(
  x,
  method = "cluster-based",
  target,
  target_buffer,
  longest_allowable_interruption = Inf,
  required_percent = 100,
  max_n_interruptions = Inf
)
}
\arguments{
\item{x}{A vector of activity labels, e.g. minute-by-minute classifications
of \code{sedentary behavior}, \code{light physical activity}, or
\code{moderate-to-vigorous physical activity}}

\item{method}{character. The bout identification method to use. Currently,
only one option is supported (\code{cluster-based}), but others may be
added in the future}

\item{target}{character. The value of \code{x} for which bout information is
desired.}

\item{target_buffer}{numeric. Maximum separation between runs of
\code{target}, beyond which they will not be clustered together}

\item{longest_allowable_interruption}{numeric. The number of consecutive
out-of-class readings that must occur to satisfy the bout termination
criteria}

\item{required_percent}{numeric (1-100). The minimum percentage of the full
bout period that must be spent engaged in the target behavior. Stated
differently, interruptions can compose no more than
(100-\code{required_percent})\% of the bout.}

\item{max_n_interruptions}{numeric. The maximum number of interruption events
that are allowed before a bout will be considered invalid.}
}
\value{
data frame with bout information
}
\description{
Identify bouts of activity, allowing for interruptions
}
\note{
\code{x} must be an atomic vector. Factors should be cast to integer or
  character prior to passing into this function.
}
\examples{
data(ex_data, package = "PAutilities")
intensity <- as.character(get_intensity(ex_data$METs))
\donttest{
get_bouts(intensity, "cluster-based", "MVPA", 30, 5, 50, 3)
}
}
\seealso{
\code{\link{bout_expand}}
}
