% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matchings.R
\name{get_matchings}
\alias{get_matchings}
\title{Obtain the matchings for predicted and actual activity transitions using the
college admissions algorithm}
\usage{
get_matchings(prefs)
}
\arguments{
\item{prefs}{an object passed from \code{\link{get_preferences}}}
}
\value{
A data frame giving the relative and absolute indices of the
  matchings, based on optimal outcomes for the students (i.e., the actual
  transitions)
}
\description{
Obtain the matchings for predicted and actual activity transitions using the
college admissions algorithm
}
\keyword{internal}
