% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ree.R
\name{get_ree}
\alias{get_ree}
\alias{get_ree_dataframe}
\alias{get_ree_default}
\title{Calculate resting energy expenditure}
\usage{
get_ree(
  method = c("harris_benedict", "schofield_wt", "schofield_wt_ht", "fao",
    "muller_wt_ht", "muller_ffm"),
  sex,
  age_yr = NULL,
  ...,
  df = NULL
)

get_ree_dataframe(df, method, sex, age_yr, ...)

get_ree_default(method, sex, age_yr, ...)
}
\arguments{
\item{method}{character. The equation to use, one of \code{"harris_benedict",
"schofield_wt", "schofield_wt_ht", "fao", "muller_wt_ht", or "muller_ffm"}}

\item{sex}{character. The participant/patient sex, one of \code{"female" or
"male"}.}

\item{age_yr}{numeric. The participant age in years. Does not need to be
passed for \code{method = "muller_ffm"}}

\item{...}{arguments (e.g. \code{wt_kg} or \code{ht_cm}) for calculations. An
error message will clarify which variables need to be passed if they are
missing.}

\item{df}{optional data frame. If passed, all prior arguments should be
character scalars pointing to a column in \code{df} that contains the
corresponding information is stored}
}
\value{
Calculated resting energy expenditure
}
\description{
Calculate resting energy expenditure
}
\examples{
get_ree("schofield_wt_ht", "female", 57.8, wt_kg = 80, ht_m = 1.50)
}
\keyword{internal}
