% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ree_helpers.R
\name{get_ree_internal}
\alias{get_ree_internal}
\alias{ree_single}
\alias{get_stratum}
\title{Internal functions for calculating resting energy expenditure}
\usage{
ree_single(
  sex = c("male", "female"),
  age_yr,
  method,
  breaks = c(-Inf, 3, 10, 18, 30, 60, Inf),
  labels = c("less3", "3to10", "10to18", "18to30", "30to60", "over60"),
  ...
)

get_stratum(method, sex, age_yr, breaks, labels)
}
\arguments{
\item{sex}{character. The participant/patient sex, one of \code{"female" or
"male"}.}

\item{age_yr}{numeric. The participant age in years. Does not need to be
passed for \code{method = "muller_ffm"}}

\item{method}{character. The equation to use, one of \code{"harris_benedict",
"schofield_wt", "schofield_wt_ht", "fao", "muller_wt_ht", or "muller_ffm"}}

\item{breaks}{age breaks to use (passed to \code{base::cut})}

\item{labels}{age labels to use (passed to \code{base::cut})}

\item{...}{arguments (e.g. \code{wt_kg} or \code{ht_cm}) for calculations. An
error message will clarify which variables need to be passed if they are
missing.}
}
\description{
Internal functions for calculating resting energy expenditure
}
\keyword{internal}
