% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metabolic_Equations.R
\name{get_schofield}
\alias{get_schofield}
\title{Retrieve Schofield basal metabolic rate for an individual}
\usage{
get_schofield(Sex, Ht = NULL, Wt, Age, verbose = TRUE,
  equation = c("both", "ht_wt", "wt"), MJ_conversion = 239.006,
  kcal_conversion = 4.86)
}
\arguments{
\item{Sex}{The participant's sex, either \code{M} or \code{F}}

\item{Ht}{The participant's height, in meters}

\item{Wt}{The participant's weight, in kilograms}

\item{Age}{The participant's age, in years}

\item{verbose}{Logical. Should processing updates be printed?}

\item{equation}{The equation to apply}

\item{MJ_conversion}{The value to use for kcals per megajoule}

\item{kcal_conversion}{The value to use for kcals per L of oxygen consumption}
}
\description{
Retrieve Schofield basal metabolic rate for an individual
}
