% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vsa.R
\name{get_vsa}
\alias{get_vsa}
\alias{get_vsa.default}
\alias{get_vsa.data.frame}
\title{Calculate body volume, surface area, or similar}
\usage{
get_vsa(x = NULL, method = c("Sendroy1966"), ...)

\method{get_vsa}{default}(x = NULL, method = c("Sendroy1966"), sex, ht, wt, ...)

\method{get_vsa}{data.frame}(x = NULL, method = c("Sendroy1966"), sex, ht, wt, ...)
}
\arguments{
\item{x}{object to operate on}

\item{method}{character. The method to use for calculations. Currently only
one method is supported (see \code{\link{sendroy_vsa}})}

\item{...}{arguments passed to methods}

\item{sex}{character. For default method, one of \code{female} or
\code{male}. For data frame method, a column name in \code{x} giving a
vector with all elements in \code{c("female", "male")}}

\item{ht}{For default method, a numeric scalar giving height in centimeters.
For data frame method, a column name in \code{x} giving a vector of heights
in centimeters.}

\item{wt}{For default method, a numeric scalar giving weight in kilograms.
For data frame method, a column name in \code{x} giving a vector of weights
in kilograms}
}
\value{
The body volume, surface area, or similar desired output
}
\description{
Calculate body volume, surface area, or similar
}
\examples{
df <- data.frame(
  sex = c("female", "male"),
  ht = c(138, 138),
  wt = c(57, 57)
)

get_vsa(sex = "female", ht = 138, wt = 57)
get_vsa(sex = "male", ht = 138, wt = 57)

get_vsa(df, sex = "sex", ht = "ht", wt = "wt")
}
