% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{mean_sd}
\alias{mean_sd}
\title{Compute the mean and standard deviation of a vector, returning a formatted
string containing the values as `M +/- SD`}
\usage{
mean_sd(x = NULL, MoreArgs = NULL, give_df = TRUE, ..., mean_x = NULL,
  sd_x = NULL)
}
\arguments{
\item{x}{numeric vector of values to summarize}

\item{MoreArgs}{named list of arguments to pass to \code{mean} and \code{sd}}

\item{give_df}{logical. Should mean, sd, and summary string be returned in a
data frame?}

\item{...}{additional arguments passed to \code{format}}

\item{mean_x}{an already-calculated mean value for \code{x}}

\item{sd_x}{an already-calculated sd value for \code{x}}
}
\description{
Compute the mean and standard deviation of a vector, returning a formatted
string containing the values as `M +/- SD`
}
\examples{
mean_sd(rnorm(1:100, 50))

}
