% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic_row_wise.R
\name{metabolic_row_wise}
\alias{metabolic_row_wise}
\alias{assemble_wt_ht}
\alias{assemble_wt}
\alias{assemble_fao}
\title{Dispatch resting metabolic rate calculations based on data frame input that
specifies conversions etc.}
\usage{
metabolic_row_wise(calculation, weights)

assemble_wt_ht(calculation, weights)

assemble_wt(calculation, weights)

assemble_fao(calculation, weights)
}
\arguments{
\item{calculation}{data frame giving instructions and values for the
calculations}

\item{weights}{list of weights to choose from for the calculation}
}
\description{
Dispatch resting metabolic rate calculations based on data frame input that
specifies conversions etc.
}
\keyword{internal}
