% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matchings.R
\name{n_crossings}
\alias{n_crossings}
\title{Determine the number of temporal conflicts among the pairs assigned by the
Gale-Shapley algorithm}
\usage{
n_crossings(matchings)
}
\arguments{
\item{matchings}{data frame with pairing information.}
}
\value{
integer vector specifying the number of conflicts for each pairing
}
\description{
Determine the number of temporal conflicts among the pairs assigned by the
Gale-Shapley algorithm
}
\keyword{internal}
