% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spurious_curve.R
\name{plot.spurious_curve}
\alias{plot.spurious_curve}
\title{Plot a spurious curve}
\usage{
\method{plot}{spurious_curve}(x, ...)
}
\arguments{
\item{x}{a \code{spurious_curve} object}

\item{...}{further arguments (currently unused)}
}
\value{
a plot of the object
}
\description{
Plot a spurious curve
}
\examples{
set.seed(100)
predictions <- (sample(1:100)\%\%2)
references  <- (sample(1:100)\%\%2)
\donttest{
trans <- get_transition_info(
  predictions, references, 7
)
result <- spurious_curve(trans)
plot(result)
}
}
\seealso{
\code{\link{spurious_curve}}
}
