% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_wearable_ActiGraph.R,
%   R/process_wearable_Generic.R
\name{process_wearable.ActiGraph}
\alias{process_wearable.ActiGraph}
\alias{process_wearable}
\title{Process data from ActiGraph monitors}
\usage{
\method{process_wearable}{ActiGraph}(file = NULL, RAW = NULL, IMU = NULL,
  TwoRegression = NULL, CutPoint = NULL, Sojourn = NULL,
  Count_Args = NULL, RAW_Args = NULL, IMU_Args = NULL, ...)

process_wearable(file = NULL, RAW = NULL, IMU = NULL,
  TwoRegression = NULL, CutPoint = NULL, Sojourn = NULL, ...,
  device = c("ActiGraph", "GENEactiv", "Axivity"))
}
\arguments{
\item{file}{character scalar giving path to an activity counts data file}

\item{RAW}{character scalar giving path to a raw primary accelerometer data
file}

\item{IMU}{character scalar giving path to an inertial measurement unit data
file}

\item{TwoRegression}{optional. One or more two-regression algorithms to apply
from \code{\link[TwoRegression]{TwoRegression-package}}}

\item{CutPoint}{optional. One or more cut-points to apply}

\item{Sojourn}{optional. One or more Sojourn methods to apply}

\item{Count_Args}{optional. named vector of arguments to be passed to
\code{\link[AGread]{read_AG_counts}}}

\item{RAW_Args}{optional. named vector of arguments to be passed to
\code{\link[AGread]{read_AG_raw}}}

\item{IMU_Args}{optional. named vector of arguments to be passed to
\code{\link[AGread]{read_AG_IMU}}}

\item{...}{further arguments passed to methods}

\item{device}{character scalar defining which device generated the data in
the file}
}
\value{
A data frame of processed data from the specified ActiGraph file(s)

A data frame of processed data from the specified monitor data file(s).
}
\description{
Process data from ActiGraph monitors

Process data obtained from a wearable device
}
\note{
Processing is currently only supported for files in .csv format.
}
\examples{
\dontrun{
# Real examples will be necessary
}
\dontrun{
# Insert a real example
}
}
