% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_transitions.R
\name{reconstruct_transitions}
\alias{reconstruct_transitions}
\title{Impute placeholder values into a \code{transition} object}
\usage{
reconstruct_transitions(x)
}
\arguments{
\item{x}{Effectively a \code{transition} object (passed before class
assignment)}
}
\description{
If missing values are passed to \code{\link{get_transition_info}}, the
prediction and criterion data streams become irregularly spaced. This fixes
the spacing issue.
}
\keyword{internal}
