% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reintegrate.R
\name{reintegrate}
\alias{reintegrate}
\title{Reintegrate accelerometer data}
\usage{
reintegrate(df, target_sec, time_var = "Timestamp")
}
\arguments{
\item{df}{a data frame to reintegrate}

\item{target_sec}{the desired epoch length of the output. Starting epoch
length will be determined automatically}

\item{time_var}{The name of the column containing POSIX-formatted timestamp
information}
}
\value{
A data frame of reintegrated data. Numeric variables will be summed,
  and others will be labeled with the first value.
}
\description{
Reintegrate accelerometer data
}
\examples{

##Example data in 60-second epochs
data(ex_data, package = "PAutilities")
ex_data$Timestamp <- as.POSIXct(
  paste(ex_data$Date, ex_data$Time),
  tz = "UTC",
  format = "\%m/\%d/\%Y \%H:\%M:\%S"
)

##Reintegrate to 120-second epochs
result <- reintegrate(ex_data, 120)

utils::head(result)

}
