% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vsa.R
\name{sendroy_vsa}
\alias{sendroy_vsa}
\title{Calculate volume-to-surface-area ratio from height and weight}
\usage{
sendroy_vsa(sex = c("female", "male"), ht, wt)
}
\arguments{
\item{sex}{character. The sex of the individual}

\item{ht}{numeric. The individual's height in centimeters}

\item{wt}{numeric. The individual's weight in kilograms}
}
\value{
The volume-to-surface-area ratio
}
\description{
Calculate volume-to-surface-area ratio from height and weight
}
\seealso{
\href{https://journals.physiology.org/doi/pdf/10.1152/jappl.1966.21.1.167}{Sendroy et al. (1966)}
}
\keyword{internal}
