% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matchings.R
\name{sequence_check}
\alias{sequence_check}
\title{Label matchings for rejection if the assignment violates sequential nature of
the data}
\usage{
sequence_check(prefs)
}
\arguments{
\item{prefs}{a list assembled in \code{\link{get_transition_info}} prior to
assignment as class \code{transition}}
}
\value{
a list ready for assignment as class \code{transition}
}
\description{
Label matchings for rejection if the assignment violates sequential nature of
the data
}
\keyword{internal}
