% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_transition.R
\docType{methods}
\name{subtract_summaryTransition}
\alias{subtract_summaryTransition}
\alias{-,summaryTransition,summaryTransition-method}
\title{Subtraction for objects of class \code{summaryTransition}}
\usage{
subtract_summaryTransition(e1, e2)

\S4method{-}{summaryTransition,summaryTransition}(e1, e2)
}
\arguments{
\item{e1}{the first object}

\item{e2}{the second object}
}
\description{
Subtraction for objects of class \code{summaryTransition}
}
\keyword{internal}
