% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_transition.R
\docType{class}
\name{summaryTransition-class}
\alias{summaryTransition-class}
\alias{summaryTransition}
\title{An S4 class containing summary information about a \code{transition} object}
\description{
An S4 class containing summary information about a \code{transition} object
}
\section{Slots}{

\describe{
\item{\code{result}}{a data frame with the summary information}
}}

