% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_mvpa.R
\name{bout_mvpa}
\alias{bout_mvpa}
\title{Classify moderate-to-vigorous physical activity in bouts of a specific
minimum length}
\usage{
bout_mvpa(intensity, var_type = c("METs", "Intensity"), min_duration = 10,
  termination = 3, MoreArgs = list(breaks = c(-Inf, 1.51, 3, Inf), labels =
  c("SB", "LPA", "MVPA"), right = FALSE), ...)
}
\arguments{
\item{intensity}{a vector of intensity classifications to be re-classified
according to the bout definition}

\item{var_type}{character scalar indicating whether the \code{intensity}
variable is a numeric vector of metabolic equivalents, or a factor variable
giving activity intensity classification}

\item{min_duration}{numeric scalar: minimum duration of a qualifying bout, in
minutes}

\item{termination}{numeric scalar: consecutive minutes of non-MVPA required
to terminate the bout}

\item{MoreArgs}{required arguments to pass to \code{cut}}

\item{...}{optional arguments passed to \code{cut} for converting METs to
intensity classification}
}
\description{
Classify moderate-to-vigorous physical activity in bouts of a specific
minimum length
}
\examples{
data(ex_data, package = "PAutilities")
bout_mvpa(intensity = ex_data$METs, var_type = "METs")

}
