% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_general_1Wrappers.R,
%   R/bout_general_2ClusterBased.R
\name{bouts_internal}
\alias{bouts_internal}
\alias{collapse_runs}
\alias{group_runs_clusterBased}
\alias{process_clusterBased_set}
\alias{check_clusterBased}
\alias{test_clusterBased_bouts}
\title{Bout analysis internal functions}
\usage{
collapse_runs(x)

group_runs_clusterBased(
  x,
  target,
  target_buffer,
  required_percent,
  longest_allowable_interruption,
  max_n_interruptions
)

process_clusterBased_set(
  runs,
  x,
  required_percent,
  longest_allowable_interruption,
  max_n_interruptions
)

check_clusterBased(x, target = NULL, required_percent, target_buffer)

test_clusterBased_bouts(
  x,
  runs,
  tree,
  h,
  required_percent,
  max_n_interruptions,
  longest_allowable_interruption
)
}
\arguments{
\item{x}{A vector of activity labels, e.g. minute-by-minute classifications
of \code{sedentary behavior}, \code{light physical activity}, or
\code{moderate-to-vigorous physical activity}}

\item{target}{character. The value of \code{x} for which bout information is
desired.}

\item{target_buffer}{numeric. Maximum separation between runs of
\code{target}, beyond which they will not be clustered together}

\item{required_percent}{numeric (1-100). The minimum percentage of the full
bout period that must be spent engaged in the target behavior. Stated
differently, interruptions can compose no more than
(100-\code{required_percent})\% of the bout.}

\item{longest_allowable_interruption}{numeric. The number of consecutive
out-of-class readings that must occur to satisfy the bout termination
criteria}

\item{max_n_interruptions}{numeric. The maximum number of interruption events
that are allowed before a bout will be considered invalid.}

\item{runs}{pre-processed input (mostly run length encoded)}
}
\description{
Bout analysis internal functions
}
\keyword{internal}
