% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_general_1Wrappers.R,
%   R/bout_general_2Sequential.R, R/bout_general_3Targeted.R
\name{bouts_internal}
\alias{bouts_internal}
\alias{collapse_runs}
\alias{group_runs_sequential}
\alias{group_runs_targeted}
\alias{check_targeted}
\alias{test_targeted_bouts}
\title{Bout analysis internal functions}
\usage{
collapse_runs(x)

group_runs_sequential(x, longest_allowable_interruption)

group_runs_targeted(
  x,
  target,
  required_percent,
  longest_allowable_interruption,
  max_n_interruptions
)

check_targeted(x, target = NULL, required_percent)

test_targeted_bouts(
  x,
  runs,
  tree,
  h,
  required_percent,
  max_n_interruptions,
  longest_allowable_interruption
)
}
\arguments{
\item{x}{A vector of activity labels, e.g. minute-by-minute classifications
of \code{sedentary behavior}, \code{light physical activity}, or
\code{moderate-to-vigorous physical activity}}

\item{longest_allowable_interruption}{numeric. The number of consecutive
out-of-class readings that must occur to satisfy the bout termination
criteria}

\item{target}{character. The value of \code{x} for which bout information is
desired (applicable only when \code{method = "targeted"})}

\item{required_percent}{numeric (1-100). The minimum percentage of time
within the bout period that must be non-interrupted in order for the bout
to qualify as valid. See details.}

\item{max_n_interruptions}{numeric. The maximum number of interruption events
that are allowed before a bout will be considered invalid.}
}
\description{
Bout analysis internal functions
}
\keyword{internal}
