% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_procedure.R
\name{manage_procedure}
\alias{manage_procedure}
\alias{get_duration}
\title{Printing and timing utility for managing processes}
\usage{
manage_procedure(part = c("Start", "End"), ..., verbose = TRUE)

get_duration(timer)
}
\arguments{
\item{part}{character scalar, either \code{Start} or \code{End}.}

\item{...}{character strings to print}

\item{verbose}{logical. Print to console?}

\item{timer}{a proc_time object}
}
\value{
For \code{part = "Start"}, a proc_time object; for \code{pard =
  "End"}, invisible
}
\description{
Printing and timing utility for managing processes
}
\examples{

timer <- manage_procedure("Start", "String will be printed")
get_duration(timer)

manage_procedure("Start")
manage_procedure("Start", "String will not be printed", verbose = FALSE)

manage_procedure(
  "End",
  "Processing complete. Elapsed time",
  get_duration(timer),
  "minutes."
)
}
