% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAutilities.R
\docType{package}
\name{PAutilities}
\alias{PAutilities}
\alias{PAutilities-package}
\title{PAutilities: Utility functions for physical activity research}
\description{
A collection of utilities that are useful for a broad range of tasks that are
common in physical activity research.
}
