% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transition.R
\name{plot.transition}
\alias{plot.transition}
\title{Plot the transitions and matchings from a \code{transition} object}
\usage{
\method{plot}{transition}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{further methods passed to or from methods, currently unused}
}
\value{
A plot of the predicted and actual transitions in a \code{transition}
  object, as well as the matchings between them
}
\description{
Plot the transitions and matchings from a \code{transition} object
}
\examples{
predictions <- sample(c(0,1), 100, TRUE, c(3, 1))
references  <- sample(c(0,1), 100, TRUE, c(4,1))
transitions <- get_transition_info(predictions, references, 10)
plot(transitions)
}
